Filter volumes by type (rw) and relationship
Added call to retrieve volumes, querying for volumes
with read/write enabled.

Added call to retrieve relationships and filtered
for the volume path client-side.

Once the path is obtained, the volume name is parsed
and the volumes that do not appear in the second call
are not in relationships, making them eligible for the
user to select during their snaplock workflow.

TODO: figure out why source/destination volumes returned
in relationship call do not appear in call for all
volumes.